+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Unreleased in the current development version (target v0.24.0):
 
 Complete list:
 - GlobalBiases: add target spacing option for stippling (#305)
+- Global Biases: fix titles and descriptions (#284)
 - CI/CD: improve temporary folders handling (#302)
 - LatLonProfiles: seasonal profiles computation fix (#288)
 - Environment LUMI dependencies update (#297)